Corrected database initialization

+++ qml/js/storage.js
@@ -28,8 +28,8 @@ function getDatabase() {
 function initDatabase() {
     console.log("initDatabase()")
     var db = getDatabase()
-    if (db.version !== "3") {
-        db.changeVersion("2", "3", function(tx) {
+    if (db.version !== "" && db.version !== "3") {
+        db.changeVersion(db.version, "3", function(tx) {
             console.log("... recreate tables")
             tx.executeSql("DROP TABLE settings")
         })
